requirements.txt
spacy
pandas
bs4
numpy
python-docx